Github's parser automatically replaces $$$$ in markdown files :unamused:
Worse, it's set up so that without an empty line before it, it replaces
it with inline math \(\) instead of display math \[\]. :unamused:
Why do this if it's not a standard markdown feature, and there's no
built in support for latex rendering? It seems pointless. It's not like
Jekyll interprets html elements in a context dependent way like this.
:unamused:  :unamused:  :unamused: